backend/requirements.txt
FastAPI
uvicorn
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
supabase
pytest
pytest-asyncio # Added
pydantic[email]
google-generativeai